requirements.txt
flask
cohere
requests
pydantic
pillow
numpy
python-dotenv